# docs/swagger.yaml
definitions:
...
  go-enterprise-api_internal_domain_permission.Permission:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  go-enterprise-api_internal_domain_ratelimit.Rule:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      requests:
        type: integer
      scope:
        type: string
      updated_at:
        type: string
      window_seconds:
        type: integer
    type: object
...
  go-enterprise-api_internal_interface_httpapi.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  go-enterprise-api_internal_interface_httpapi.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorBody'
    type: object
...
  internal_interface_http_handler.permissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  internal_interface_http_handler.ruleRequest:
    properties:
      enabled:
        type: boolean
      identifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      requests:
        type: integer
      scope:
        type: string
      window_seconds:
        type: integer
    type: object
...
host: localhost:18080
info:
  contact:
    name: GitOps
  description: 'Enterprise boilerplate: Dynamic Rate Limit, Soft Delete, Audit Logs,
    RBAC, JWT + Redis sessions, Microsoft Entra (OIDC). All requests pass through
    the Nginx gateway (X-Gateway-Token handshake) — public endpoints are also rate-limited
    per IP at the gateway.'
  title: Go Enterprise API
  version: "1.0"
paths:
...
  /api/v1/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_permission.Permission'
            type: array
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      parameters:
      - description: Permission
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.permissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_permission.Permission'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create permission
      tags:
      - permissions
  /api/v1/permissions/{id}:
    delete:
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete permission
      tags:
      - permissions
    put:
      consumes:
      - application/json
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.permissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_permission.Permission'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - permissions
  /api/v1/permissions/{id}/restore:
    post:
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore soft-deleted permission
      tags:
      - permissions
  /api/v1/rate-limits:
    get:
      description: Dynamic rules stored in PostgreSQL. Changes apply live within the
        Redis cache TTL (30s).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_ratelimit.Rule'
            type: array
      security:
      - BearerAuth: []
      summary: List rate limit rules
      tags:
      - rate-limits
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.ruleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_ratelimit.Rule'
      security:
      - BearerAuth: []
      summary: Create rate limit rule
      tags:
      - rate-limits
  /api/v1/rate-limits/{id}:
    delete:
      description: Disables the rule.
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete rate limit rule
      tags:
      - rate-limits
    put:
      consumes:
      - application/json
      description: 'Example: raise the Viewer role from 500 to 1000 req/min without
        redeploying. Audited old → new.'
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rule
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.ruleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_ratelimit.Rule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update rate limit rule
      tags:
      - rate-limits
  /api/v1/rate-limits/{id}/restore:
    post:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore soft-deleted rate limit rule
      tags:
      - rate-limits
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Provide a JWT access token obtained from /api/v1/auth/login or the
      Entra OIDC flow as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"